added props to add user callbacks to UI events
In LoginForm component, added props

- onSignOut
- onSwitchToSignUp
- onSwitchToSignIn
- onSwitchToPasswordReset
- onSwitchToChangePassword
- onSignUp
- onSignIn
- onChangePassword
- onSwitchToSignOut

to make it possible for developers to add their own callbacks to
click events for corresponding basic methods:

- signOut
- switchToSignUp
- switchToSignIn
- switchToPasswordReset
- switchToChangePassword
- signUp
- signIn
- changePassword
- switchToSignOut

in a form of

  onClick: function (event) {
    if (_.isFunction(this.props.onX)) {
      this.props.onX(event);
    }

    this.X(event);
  }

Each of these methods is a function assigned to onClick of a
corresponding button DOM element. Therefore, it's now possible to
do something like

  <Accounts.ui.LoginForm
    onSignIn={() => console.log('clicked "Sign In" button!')} />

and receive this message in the console every time a user clicks
"Sign In" button.